Fix pkg.install when using a port in the url.
