+++ changelog/64516.fixed.md
@@ -1,0 +1,1 @@
+Fix pkg.install when using a port in the url.
